pyproject.toml
[project]
# Project metadata. Available keys are documented at:
# https://packaging.python.org/en/latest/specifications/declaring-project-metadata
name = 'dataset_grouper'
description = 'Dataset Grouper - A library for datasets with group-level structure.'
readme = 'README.md'
requires-python = '>=3.10'
license = {file = 'LICENSE'}
authors = [{name = 'Dataset Grouper Team', email='dataset-grouper-dev@google.com'}]
classifiers = [  # List of https://pypi.org/classifiers/
    'License :: OSI Approved :: Apache Software License',
    'Intended Audience :: Science/Research',
]
keywords = ['federated learning', 'datasets', 'machine learning']

# pip dependencies of the project
# Installed locally with `pip install -e .`
dependencies = [
  'absl-py~=1.4.0',
  'apache-beam>=2.47.0',
  'numpy~=1.21',
  'tensorflow~=2.12.0',
  'tensorflow-datasets~=4.9.0',
]

# `version` is automatically set by flit to use `dataset_grouper.__version__`
dynamic = ['version']

[build-system]
# PEP 508 specs
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
# Development deps (unittest, linting, formating,...)
# Installed through `pip install -e .[dev]`
dev = [
    "pytest",
    "pytest-xdist",
]

[project.urls]
homepage = 'https://github.com/google-research/dataset_grouper'
repository = 'https://github.com/google-research/dataset_grouper'
changelog = 'https://github.com/google-research/dataset_grouper/blob/main/CHANGELOG.md'
